⚡Core: Server Page JS entrypoint refactoring
